Deploy

+++ spider/spider.py
@@ -1,23 +1,22 @@
 import os
+import sys
 import re
 import shutil
 import string
-import sys
 import time
 import traceback
 import urllib.error
 import urllib.parse
 import urllib.request
 from http.client import HTTPResponse
-from queue import Queue
-from typing import Set
 from urllib.parse import quote
 
 from bs4 import BeautifulSoup
+import selenium
 from selenium import webdriver
 from selenium.webdriver import FirefoxProfile
-from selenium.webdriver.firefox.options import Options as FireFoxOptions
 from selenium.webdriver.chrome.options import Options as ChromeOptions
+from selenium.webdriver.firefox.options import Options as FireFoxOptions
 
 from myparser import MyHTMLParser
 from mytool import Config, MyUtil
@@ -31,7 +30,7 @@ class Spider:
     doc_folder = Config.get("path.document")
     download_temp_folder = Config.get("path.download_temp_dir")
 
-    def __init__(self, debug_mode=False, download_file=True):
+    def __init__(self, download_file=True, debug_mode=False):
         # self.bfs_queue = Queue()
         # self.visited: Set[str] = set()
         MyUtil.create_folders()
@@ -45,13 +44,21 @@ class Spider:
                 self.profile.set_preference("browser.download.manager.showWhenStarting", False)
                 self.profile.set_preference("browser.download.dir", self.download_temp_folder)
                 self.profile.set_preference("browser.helperApps.neverAsk.saveToDisk", "application/octet-stream")
+            self.profile.set_preference("permissions.default.image", 2)  # no images
+            self.profile.set_preference("permissions.default.stylesheet", 2)  # no CSS
+            self.profile.set_preference("dom.ipc.plugins.enabled.libflashplayer.so", False)  # no Flash
             self.driver_options = FireFoxOptions()
             self.driver_options.add_argument('--headless')
         elif Config.get("spider.browser") == "chrome":
             self.driver_options = ChromeOptions()
-            self.driver_options.add_argument('--headless')
+            # self.driver_options.add_argument('--no-sandbox')
+            # self.driver_options.add_argument('--disable-dev-shm-usage')
+            # self.driver_options.add_argument('--disable-gpu')
+            # self.driver_options.add_argument('--log-level=3')
+            # self.driver_options.add_argument('--headless')
             prefs = {'profile.default_content_settings.popups': 0,
-                     'download.default_directory': self.download_temp_folder}
+                     'download.default_directory': self.download_temp_folder,
+                     "profile.managed_default_content_settings.images":2}  # no images
             self.driver_options.add_experimental_option('prefs', prefs)
         self.driver = None
 
@@ -110,22 +117,22 @@ class Spider:
         # print("!")
         doc_path = self.doc_folder + MyUtil.md5(url) + ".txt"
         page_path = self.page_folder + MyUtil.md5(url) + ".html"
-        if not os.path.exists(doc_path) or not MyRedisUtil.is_visited(url):
-            MyRedisUtil.first_time(url)
+        if not os.path.exists(doc_path) or not MyRedisUtil.have_visited(url):
+            MyRedisUtil.first_time_visit(url)
             self.write_page(html_text, page_path)
             self.write_doc(parser.text, doc_path)
         else:
             old_text = MyUtil.read_str(doc_path)
             diff_ratio = MyUtil.diff_ratio(old_text, parser.text)
             if diff_ratio > 0.99:
-                MyRedisUtil.unchanged(url)
+                MyRedisUtil.unchanged_visit(url)
             else:
-                MyRedisUtil.changed(url)
+                MyRedisUtil.changed_visit(url)
                 self.write_page(html_text, page_path)
                 self.write_doc(parser.text, doc_path)
 
     def process_html_by_webdriver(self, url: str):
-        # print("process_html_by_webdriver")
+        print("process_html_by_webdriver", file=self.log_file)
         if self.driver is None:
             if Config.get("spider.browser") == "firefox":
                 self.driver = webdriver.Firefox(firefox_profile=self.profile, options=self.driver_options)
@@ -135,20 +142,30 @@ class Spider:
             self.driver.set_page_load_timeout(Config.get("spider.driver.page_load_timeout"))  # in seconds
             self.driver.set_script_timeout(Config.get("spider.driver.script_timeout"))  # in seconds
 
+        if self.download_file:
+            for file_name in os.listdir(self.download_temp_folder):
+                os.remove(self.download_temp_folder + "\\" + file_name)
+
         self.driver.get(url)
+        try:
+            self.driver.switch_to.alert.accept()
+        except selenium.common.exceptions.NoAlertPresentException:
+            pass
+        except:
+            raise
+        
         html_text = self.driver.page_source
         self.process_html_text(html_text, url)
 
-        for file_name in os.listdir(self.download_temp_folder):
-            os.remove(self.download_temp_folder + "\\" + file_name)
-        time.sleep(Config.get("spider.wait_download_max_sec"))
-        if os.listdir(self.download_temp_folder):
-            file_name = os.listdir(self.download_temp_folder)[0]
-            old_path = self.download_temp_folder + "\\" + file_name
-            ext = file_name.split(".")[-1] if "." in file_name else ""
-            new_file_name = MyUtil.md5(url) + "." + ext
-            new_path = self.doc_folder + "\\" + new_file_name
-            shutil.move(old_path, new_path)
+        if self.download_file:
+            time.sleep(Config.get("spider.driver.wait_download_max_sec"))
+            if os.listdir(self.download_temp_folder):
+                file_name = os.listdir(self.download_temp_folder)[0]
+                old_path = self.download_temp_folder + "\\" + file_name
+                ext = file_name.split(".")[-1] if "." in file_name else ""
+                new_file_name = MyUtil.md5(url) + "." + ext
+                new_path = self.doc_folder + "\\" + new_file_name
+                shutil.move(old_path, new_path)
 
     def process_html(self, response: HTTPResponse, url: str) -> bool:
         # print("process_html")
@@ -179,9 +196,9 @@ class Spider:
             self.process_file(response, url)
 
     def search(self, url: str):
-        print(time.time(), "searching: ", url, file=self.log_file)
+        print(time.time(), "searching: ", url, MyUtil.md5(url), file=self.log_file)
         # self.visited.add(url)
-        MyRedisUtil.set_visited(url)
+        MyRedisUtil.push_visited(url)
         if not MyRedisUtil.need_search(url) or not self.url_validation(url):
             return
 
@@ -189,25 +206,36 @@ class Spider:
             self.process_one_url(url)
         except (urllib.error.HTTPError, urllib.error.URLError, ConnectionResetError) as error:
             MyRedisUtil.set_known_exception(url, error)
-            MyRedisUtil.exceptional(url)
+            MyRedisUtil.exceptional_visit(url)
             print("[exception]", type(error), error, file=self.log_file)
+        except selenium.common.exceptions.WebDriverException as error:
+            self.driver.quit()
+            self.driver = None
+            MyRedisUtil.set_known_exception(url, error)
+            MyRedisUtil.exceptional_visit(url)
+            print("[exception]", type(error), error, file=self.log_file)
+        except (KeyboardInterrupt, SystemExit):
+            raise
         except BaseException as error:
             MyRedisUtil.set_unknown_exception(url, error)
-            MyRedisUtil.exceptional(url)
+            MyRedisUtil.exceptional_visit(url)
             print("[exception]", type(error), error, file=self.log_file)
             traceback.print_exc(file=self.log_file)
 
-    def new_job(self):
+    @staticmethod
+    def new_job():
         MyRedisUtil.flush()
         # self.bfs_queue.put(Config.get("job.start_url"))
         MyRedisUtil.push_need_search(Config.get("job.start_url"))
 
-    def resume_batch(self):
+    @staticmethod
+    def resume_batch():
         # self.bfs_queue = MyRedisUtil.get_queue()
         # self.visited = MyRedisUtil.get_visited()
         pass
 
-    def new_batch(self):
+    @staticmethod
+    def new_batch():
         all_urls = MyRedisUtil.get_all_urls()
         if all_urls:
             for url in all_urls:
@@ -222,8 +250,11 @@ class Spider:
 
         if mode == "new_job":
             self.new_job()
-        elif mode == "resume" and MyRedisUtil.stored_breakpoint():
-            self.resume_batch()
+        elif mode == "resume":
+            if MyRedisUtil.need_search_num() == 0:
+                return False
+            else:
+                self.resume_batch()
         elif mode == "new_batch":
             self.new_batch()
         else:
@@ -240,6 +271,8 @@ class Spider:
                 # curr_url: str = MyUtil.normalize_url(self.bfs_queue.get())
                 curr_url: str = MyUtil.normalize_url(MyRedisUtil.pop_need_search())
                 self.search(curr_url)
+            except (KeyboardInterrupt, SystemExit):
+                raise
             except BaseException as error:
                 MyRedisUtil.set_unknown_exception("", error)
                 print("[exception]", type(error), error, file=self.log_file)
@@ -247,6 +280,7 @@ class Spider:
 
         # MyRedisUtil.store_visited(self.visited)
         # MyRedisUtil.store_queue(self.bfs_queue)
+        return True
 
 
 if __name__ == "__main__":
@@ -254,15 +288,16 @@ if __name__ == "__main__":
     spider = Spider(download_file=False, debug_mode=False)
 
     try:
-        mode = sys.argv[1]
-        max_doc_num = int(sys.argv[2])
+        mode_ = sys.argv[1]
+        max_doc_num_ = int(sys.argv[2])
     except:
         print("invalid parameters")
         exit(-1)
-
-    spider.run(mode, max_doc_num)
-    # spider.run("new_batch", 100)
-    # spider.run("resume", 80000)
+    
+    try:
+        spider.run(mode_, max_doc_num_)
+    except KeyboardInterrupt as error:
+        print("Exit by KeyboardInterrupt.")
 
     spider.quit()
     print("end")
